Move sync cron worker to sub directory
